// ForExample/src/app/store/features.enum.ts
export enum FeaturesEnum {
  CONFIG = 'config',
  USERS_STATE = 'usersState',
  DEBITORS_STATE = 'debitorsState',
  CUSTOMER_CONTACTS_STATE = 'customer_contacts',
  NOTIFICATIONS = 'notification',
  CANCELLATION = 'cancellation',
  AUTH = 'auth',
  HISTORY = 'history',
  SMS_CHECK_LIST = 'smsCheckList',
  Tasks = 'tasks',
}